# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API",
        "description": "API for the Laravel E-Commerce application. Provides endpoints for browsing products, managing carts/wishlists, placing orders, and customer account management.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Application server"
        },
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List products with filtering, search, and pagination",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by keyword (name, description, SKU)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Minimum price filter",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Maximum price filter",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "price_asc",
                                "price_desc",
                                "newest",
                                "name",
                                "on_sale"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page (max 50)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 16
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{slug}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show product detail by slug",
                "operationId": "5bbb6d725346d38781b256ed5c715d62",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ProductDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Wireless Headphones"
                    },
                    "slug": {
                        "type": "string",
                        "example": "wireless-headphones"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "sale_price": {
                        "type": "number",
                        "format": "float",
                        "example": 79.99,
                        "nullable": true
                    },
                    "display_price": {
                        "type": "number",
                        "format": "float",
                        "example": 79.99
                    },
                    "on_sale": {
                        "type": "boolean",
                        "example": true
                    },
                    "image_url": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "stock": {
                        "type": "integer",
                        "example": 50
                    },
                    "in_stock": {
                        "type": "boolean",
                        "example": true
                    },
                    "category": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Electronics"
                            },
                            "slug": {
                                "type": "string",
                                "example": "electronics"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ProductDetail": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Product"
                    },
                    {
                        "properties": {
                            "description": {
                                "type": "string",
                                "example": "High-quality wireless headphones with noise cancellation."
                            },
                            "sku": {
                                "type": "string",
                                "example": "WH-1000XM4"
                            },
                            "stock": {
                                "type": "integer",
                                "example": 50
                            },
                            "images": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "url": {
                                            "type": "string",
                                            "format": "uri"
                                        },
                                        "alt": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "reviews": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Review"
                                }
                            },
                            "avg_rating": {
                                "type": "number",
                                "format": "float",
                                "example": 4.5,
                                "nullable": true
                            },
                            "review_count": {
                                "type": "integer",
                                "example": 10
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "Review": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "rating": {
                        "type": "integer",
                        "example": 5,
                        "maximum": 5,
                        "minimum": 1
                    },
                    "comment": {
                        "type": "string",
                        "example": "Great product!",
                        "nullable": true
                    },
                    "user_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-06-23"
                    }
                },
                "type": "object"
            },
...
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 16
                    },
                    "total": {
                        "type": "integer",
                        "example": 80
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token from POST /api/login or POST /api/register",
                "scheme": "bearer"
            }
        }
    },
...
}